requirements.txt
langchain==0.3.17
langchain-core==0.3.33
langchain-experimental==0.3.4
langchain-community==0.3.4
python-dotenv==1.0.1
langchain_openai==0.3.3
pydantic==2.7.4
streamlit
streamlit-chat
flask
spacy